pyproject.toml
[tool.poetry]
name = "barter"
version = "0.1.0"
description = ""
authors = ["John Smith <john@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"
django-structlog = "^4.0.1"
python-dotenv = "^0.21.1"
django-cors = "^0.1"
django-cors-headers = "^3.13.0"
django-stubs = "^1.14.0"
djangorestframework-stubs = "^1.8.0"
dj-database-url = "^1.2.0"
uvicorn = "^0.22.0"
psycopg2 = "^2.9.5"
whitenoise = "^6.3.0"
pre-commit = "^3.1.1"
ruff = "^0.0.254"
pandas = "1.5.3"
celery = "^5.2.7"
django-model-utils = "^4.3.1"
dateparser = "1.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pytest = "^7.2.1"
black = "^22.12.0"
pytest-django = "^4.5.2"
django-filter = "^23.2"
pip = "^23.1.2"
jupyter = "^1.0.0"
isort = "^5.12.0"
django-adminactions = "^2.1.0"
django-sql-explorer = {extras = ["charts"], version = "^3.2"}
dydx-v3-python = "^2.0.1"
django-crispy-forms = "^2.0"
crispy-bootstrap4 = "^2022.1"
pydantic = "^1.10.9"
datamodel-code-generator = "^0.20.0"

[tool.ruff]
ignore = ["E501", "E402"]

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
